refactor: remove scope.rs
